Fix helpers such that every function is implemented exactly once from the compiler's point of view (fully specified functions in helpers.cc, implementations of template functions in helpers.h.
